correct section definitions when last section is a RECT
we ensure that if the final section is a RECT, then we must extend it
past the centre row (to the mirrored side). this will then ensure the
section iteration in the kernel is correct

+++ proj_sats/sats.cu
@@ -281,7 +281,7 @@ int main(int argc, char **argv) {
 
   double radiusPixels = 1.1;
   if (argc >= 4)
-    radiusPixels = atoi(argv[3]);
+    radiusPixels = atof(argv[3]);
   printf(" Radius in pixels: %f\n", radiusPixels);
 
   using Tin = int32_t;  // input is small
